feat: move to dealer
