+++ src/services/dealer.ts
@@ -8,6 +8,7 @@ import { Dealer, DealerSuggestion, Entitlements } from "../types/models"
 import toQueryString from "../lib/toQueryString"
 import {
   ApiCallOptions,
+  deletePath,
   fetchPath,
   handleValidationError,
   postData,
@@ -313,6 +314,26 @@ export const putUserAccount = async ({
     return {
       tag: "success",
       result,
+    }
+  } catch (error) {
+    return handleValidationError(error, { swallowErrors: true })
+  }
+}
+
+export const deleteUser = async ({
+  options = {},
+}: {
+  options?: ApiCallOptions
+}): Promise<WithValidationError> => {
+  try {
+    await deletePath({
+      path: `users/me`,
+      options: { isAuthorizedRequest: true, ...options },
+    })
+
+    return {
+      tag: "success",
+      result: {},
     }
   } catch (error) {
     return handleValidationError(error, { swallowErrors: true })
